wpa_supplicant v2.9 - Fix "Wireless Event too big (33)" messages
The Radio is observed to generate "Wireless Event too big (33)"
kernel messages when wpa_supplicant commands the wi-fi radio to
disconnect from its AP.

One circumstance has been identified: wpa_supplicant will attempt to
set a randomized bogus SSID when it disconnects or deauthenticates the
wi-fi radio, or wpa_supplicant starts up. This SSID is 32 bytes long,
the maximum allowed length of an SSID. The purpose appears to be to
clear any possible previous association the driver may retain.

It appears that the AR6000/AR6002 driver/firmware will always have
difficulty with SSIDs of length 32. The driver in use appears to
simply increment the recorded length of a provided SSID by 1. We do not
have the source code available to explain this; it might, perhaps, be an
attempt to convert from behaviour expected of a kernel running WE > 20,
into behaviour compatible with WE <= 20.

In this case the result is that an SSID of length 32 is reported as an
apparently oversized SSID of length 33, hence the kernel message.

This change applies a patch to wpa_supplicant to reduce the length of
the randomized bogus SSID to 31 bytes. The SSID is still random, but no
longer triggers the message.

The patch applies to wpa_supplicant v2.9. The relevant behaviour was
introduced by v0.6.8, and is present in v0.6.9 which is used in the
Radio's stock firmware. That also generates these kernel messages. A
v0.6.9 patch could be prepared if required.

The underlying commits in hostapd were:

wext: really disassociate (set random SSID)
commit #b965fa7
2009/02/10

wext: Force disconnect on deauthenticate()
commit #4853d5a
2009/02/10