wpa_supplicant v2.9 - 'ifdown' prevents disconnection from existing AP
Shortly after wpa_supplicant v0.7.2, changes were introduced to support
rf_kill.

Some adverse effects of these changes have been dealt with in
commit #4bf00d1:
wpa_supplicant - 'ifdown' inhibits ability to SCAN or REASSOCIATE

A further adverse effect has been identified, which prevents DHCP
configuration operating effectively when switching from one wireless
network to another.

When a wireless interface is taken down, wpa_supplicant will place the
wireless interface into an 'INTERFACE_DISABLED' state. But it does not
disconnect the radio from the access point - the connection remains
intact.

When the interface is next brought back up, wpa_supplicant transitions
into a 'DISCONNECTED' state, although the pre-existing wireless
connection still remains active.

In this state, wpa_supplicant appears to ignore a wpa_cli 'DISCONNECT'
command (presumably because it is in the 'DISCONNECTED' state).
But the pre-existing wireless connection is still active, and we are,
now, unable to disconnect from it.

This breaks Squeezeplay's network management logic, which
relies upon a fully working and timely wpa_cli 'DISCONNECT'.

In particular, when switching to a new wireless network (using
jive.net.Networking:_ifUp), it is more than likely that the DHCP
client (udhcpc) will obtain a configuration/lease for the new network
from the old, still connected, network. This does't work well.

This change adds a network/if-down action script that disconnects the
radio from its access point, *before* taking the wireless interface
down. So, when the interface is next brought back up, it will genuinely
be disconnected. This resolves the issue noted with DHCP.

At present this is only applicable to the Squeezebox Radio (baby). The
relevant interface is eth1, this is hard coded into the rest of the
firmware.

The Squeezebox Controller (jive) and Touch (fab4) use different wi-fi
interfaces, so the script would require suitable amendment if ever
adopted for those platforms.