wpa_supplicant v2.9 - Wireless scan timeout is too long for 'baby'
This patch to wpa_supplicant has been prepared in the context of the
Squeezebox Radio which uses an Atheros AR6002 wi-fi module.

In hostap commit #180cdf45, the wireless scan timeout was increased from
5 to 10 seconds. This was to accomodate dual band cards.

This patch reverts that commit, restoring the timeout to 5 seconds.

The change was introduced after wpa_supplicant v0.7.2, and is present in
v1.0.

The change has an adverse effect on Squeezeplay's UI, which anticipates
that results will become available in about 5 seconds or so.

The consequence is that the user will regularly not see any available
wireless networks when he browses into the "Choose Network" menu, other
than those that may exist in the wpa_supplicant.conf configuration file.

Reverting back to 5 seconds does improve matters, although it is not a
complete cure.

Why a timeout should be necessary is unknown. It is possible that the
Atheros driver in use in the 'baby' does not generate "scan completed"
wireless events.