Update timezone data to 2021e - Provide bespoke recipe for SqueezeOS
This recipe is a customized version of 'tzdata', which builds the timezone
database '/usr/share/zoneinfo'. It is intended to replace the timezone
database provided by the CSL tool chain.

The recipe 'tzdata' cannot be used 'as is' for this purpose, because it does
not compile suitable timezone files for this older 32 bit OS.

It is intended to be automatically invoked by recipe
'external-csl-toolchain_2010q1-202-modified.bb', which is a modified
version of the CSL tool chain that suppresses timezone database generation.

The timezone data file used is version 2021e, and the recipe depends on
recipe 'tzcode-native_2021e', which provides the 'zic' timezone compiler.

The recipe checks that all required timezone files are built, these are
specified in file 'required_tzones.tzt'.

It also creates the required '/etc/localtime' link, which defaults to the
'Factory' timezone.