Update time zone data - modify CSL tool chain to use our new tzdata-squeezeos recipe
This change introduces a modified version of the CSL tool chain, whose purpose
is to suppress the out of date timezone data packaged up with the original
tool chain, and generate timezone data using an up to date recipe.

* external-csl-toolchain_2010q1-202-modified.bb

This is the modified tool chain. It is based on the tool chain in use,
'external-csl-toolchain_2010q1-202.bb', but adds a supplementary step to
remove the provided, out of date, timezone data.

It also incorporates a run time dependency on our new, up to date, timezone
data recipe, 'tzdata-squeezeos'.

* poky-external-csl2010q1-modified.inc

A required 'include' file. It essentially replicates the existing
'poky-external-csl2010q1.inc'.

* local.conf.sample

This file has been changed to prefer our new, modified, tool chain. But it is
only a sample file, the change must be reflected in a real 'local.conf' to
have any effect.